Remove old bee.Cluster in favor or bee.DynamicCluster, and rename DynamicCluster to Cluster
