+++ cmd/beekeeper/cmd/check_balances.go
@@ -30,7 +30,7 @@ func (c *command) initCheckBalances() *cobra.Command {
 		Short: "Executes balances check",
 		Long:  `Executes balances check.`,
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			cluster := bee.NewDynamicCluster("bee", bee.DynamicClusterOptions{
+			cluster := bee.NewCluster("bee", bee.ClusterOptions{
 				APIDomain:           c.config.GetString(optionNameAPIDomain),
 				APIInsecureTLS:      insecureTLSAPI,
 				APIScheme:           c.config.GetString(optionNameAPIScheme),
