+++ cmd/beekeeper/cmd/check_fileretrieval.go
@@ -36,7 +36,7 @@ and attempts retrieval of those files from the last node in the cluster.`,
 				return errors.New("bad parameters: upload-node-count must be less or equal to node-count")
 			}
 
-			cluster := bee.NewDynamicCluster("bee", bee.DynamicClusterOptions{
+			cluster := bee.NewCluster("bee", bee.ClusterOptions{
 				APIDomain:           c.config.GetString(optionNameAPIDomain),
 				APIInsecureTLS:      insecureTLSAPI,
 				APIScheme:           c.config.GetString(optionNameAPIScheme),
