+++ cmd/beekeeper/cmd/check_localpinning.go
@@ -23,7 +23,7 @@ func (c *command) initCheckLocalPinningChunk() *cobra.Command {
 		Long:  "Checks that a node on the cluster pins one chunk correctly.",
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 
-			cluster := bee.NewDynamicCluster("bee", bee.DynamicClusterOptions{
+			cluster := bee.NewCluster("bee", bee.ClusterOptions{
 				APIDomain:           c.config.GetString(optionNameAPIDomain),
 				APIInsecureTLS:      insecureTLSAPI,
 				APIScheme:           c.config.GetString(optionNameAPIScheme),
