+++ cmd/beekeeper/cmd/check_settlements.go
@@ -31,7 +31,7 @@ func (c *command) initCheckSettlements() *cobra.Command {
 		Short: "Executes settlements check",
 		Long:  `Executes settlements check.`,
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			cluster := bee.NewDynamicCluster("bee", bee.DynamicClusterOptions{
+			cluster := bee.NewCluster("bee", bee.ClusterOptions{
 				APIDomain:           c.config.GetString(optionNameAPIDomain),
 				APIInsecureTLS:      insecureTLSAPI,
 				APIScheme:           c.config.GetString(optionNameAPIScheme),
