+++ cmd/beekeeper/cmd/start_cluster.go
@@ -35,7 +35,7 @@ func (c *command) initStartCluster() *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) (err error) {
 			ctx := context.Background()
 
-			cluster := bee.NewDynamicCluster(clusterName, bee.DynamicClusterOptions{
+			cluster := bee.NewCluster(clusterName, bee.ClusterOptions{
 				Annotations: map[string]string{
 					"created-by":        createdBy,
 					"beekeeper/version": beekeeper.Version,
